+++ cpp/runtimes/cluster-controller-runtime/JoynrClusterControllerRuntime.cpp
@@ -112,6 +112,7 @@ static const std::string ACC_ENTRIES_FILE = "CCAccessControl.entries";
 
 JoynrClusterControllerRuntime::JoynrClusterControllerRuntime(
         std::unique_ptr<Settings> settings,
+        std::shared_ptr<IKeychain> keyChain,
         std::shared_ptr<ITransportMessageReceiver> httpMessageReceiver,
         std::shared_ptr<ITransportMessageSender> httpMessageSender,
         std::shared_ptr<ITransportMessageReceiver> mqttMessageReceiver,
@@ -153,14 +154,16 @@ JoynrClusterControllerRuntime::JoynrClusterControllerRuntime(
           ccMessageRouter(nullptr),
           aclEditor(nullptr),
           lifetimeSemaphore(0),
-          accessController(nullptr)
+          accessController(nullptr),
+          keyChain(std::move(keyChain))
 {
     initializeAllDependencies();
 }
 
 std::unique_ptr<JoynrClusterControllerRuntime> JoynrClusterControllerRuntime::create(
         std::size_t argc,
-        char* argv[])
+        char* argv[],
+        std::shared_ptr<IKeychain> keyChain)
 {
     // Object that holds all the settings
     auto settings = std::make_unique<Settings>();
@@ -201,7 +204,7 @@ std::unique_ptr<JoynrClusterControllerRuntime> JoynrClusterControllerRuntime::cr
         // Merge
         Settings::merge(currentSettings, *settings, true);
     }
-    return create(std::move(settings), discoveryEntriesFile);
+    return create(std::move(settings), discoveryEntriesFile, keyChain);
 }
 
 void JoynrClusterControllerRuntime::initializeAllDependencies()
@@ -934,9 +937,11 @@ void JoynrClusterControllerRuntime::runForever()
 
 std::unique_ptr<JoynrClusterControllerRuntime> JoynrClusterControllerRuntime::create(
         std::unique_ptr<Settings> settings,
-        const std::string& discoveryEntriesFile)
+        const std::string& discoveryEntriesFile,
+        std::shared_ptr<IKeychain> keyChain)
 {
-    auto runtime = std::make_unique<JoynrClusterControllerRuntime>(std::move(settings));
+    auto runtime = std::make_unique<JoynrClusterControllerRuntime>(
+            std::move(settings), std::move(keyChain));
 
     assert(runtime->localCapabilitiesDirectory);
     runtime->localCapabilitiesDirectory->injectGlobalCapabilitiesFromFile(discoveryEntriesFile);